Text wrapping complete along with minor improvements
Changes:

-Text class now takes a new structure TextOptions, which is taken into account when rendering
-Text class now supports word wrapping
-Shader class now has the locations of common(or standard) uniforms hardcoded into it to reduce lookup time
-StrictBatch restructured to use a class-wide VAO
--Thus the render routine for StrictBatches is also improved
-SoftBatch individual buffering to its buffers supported
-Added some static diagnostic methods for OpenGL for debugging

+++ OpenGl-2D-Renderer-Revamped/components/ui/font.cpp
@@ -70,7 +70,7 @@ unsigned short Font::GetOffsetForGlyph(
 		if (_char == m_GlyphIdentifier[i]) 
 		{
 			if (OUT_glyphIndexWithinSpriteSheet) {
-				*OUT_glyphIndexWithinSpriteSheet = i;
+				*OUT_glyphIndexWithinSpriteSheet = static_cast<unsigned short>(i);
 			}
 			return m_GlyphOffsetsFromAdjecentGlyph[i];
 		}
